Se agrega telefono a Anfitrión y enlace de whats app a mail de alta de reserva.

+++ amq-backend/src/main/java/com/amq/model/Huesped.java
@@ -32,12 +32,12 @@ public class Huesped extends Usuario implements Serializable {
 	@OneToMany(mappedBy = "huesped",cascade = CascadeType.ALL,orphanRemoval=true)
 	private List<Reserva> reservas = new ArrayList<Reserva>();
 	
-	private Integer telefono;
+	private Long telefono;
 	
 	public Huesped() {
 		super();
 	}
-	public Huesped(String email, Boolean activo, String apellido, String nombre, Integer calificacion, List<String> tokens, String pass, Boolean bloqueado, Integer telefono) {
+	public Huesped(String email, Boolean activo, String apellido, String nombre, Integer calificacion, List<String> tokens, String pass, Boolean bloqueado, Long telefono) {
 		super(email, activo, bloqueado, apellido, nombre, pass);
 		this.calificacionGlobal = calificacion;
 		this.pushTokens = tokens;
@@ -74,10 +74,10 @@ public class Huesped extends Usuario implements Serializable {
 	public void setBloqueado(Boolean bloqueado) {
 		this.bloqueado = bloqueado;
 	}
-	public Integer getTelefono() {
+	public Long getTelefono() {
 		return telefono;
 	}
-	public void setTelefono(Integer telefono) {
+	public void setTelefono(Long telefono) {
 		this.telefono = telefono;
 	}
